Fix: apparently name can be "null"
Hetzner supported added an ISO for me, but I can't pull it up using your code because the name is "null". Not sure if this is a "bug" in your code, or if it's how they added my ISO in and a bug on their end. But adding `= null` fixes it. ;-)